+++ src/act_wizard.cpp
@@ -12,8 +12,6 @@
 *  $Revision$                                                      *
 ************************************************************************ */
 
-#include "act_wizard.h"
-
 #include "administration/proxy.h"
 #include "administration/ban.h"
 #include "administration/punishments.h"
@@ -52,10 +50,8 @@
 #include "engine/db/obj_prototypes.h"
 #include "engine/olc/olc.h"
 #include "gameplay/communication/parcel.h"
-#include "administration/password.h"
 #include "administration/privilege.h"
 #include "third_party_libs/pugixml/pugixml.h"
-#include "engine/ui/color.h"
 #include "gameplay/skills/skills.h"
 #include "gameplay/magic/spells.h"
 #include "engine/network/descriptor_data.h"
@@ -132,7 +128,6 @@ void do_liblist(CharData *ch, char *argument, int cmd, int subcmd);
 //
 void do_sdemigod(CharData *ch, char *argument, int cmd, int subcmd);
 void do_unfreeze(CharData *ch, char *argument, int cmd, int subcmd);
-void do_setall(CharData *ch, char *argument, int cmd, int subcmd);
 void do_check_occupation(CharData *ch, char *argument, int cmd, int subcmd);
 void do_delete_obj(CharData *ch, char *argument, int cmd, int subcmd);
 void do_arena_restore(CharData *ch, char *argument, int cmd, int subcmd);
@@ -500,300 +495,6 @@ void do_check_occupation(CharData *ch, char *argument, int/* cmd*/, int/* subcmd
 	if (!is_found) {
 		SendMsgToChar("Такой зоны нет.\r\n", ch);
 	}
-}
-
-SetAllInspReqListType &setall_inspect_list = MUD::setall_inspect_list();
-
-const int SETALL_FREEZE = 0;
-const int SETALL_EMAIL = 1;
-const int SETALL_PSWD = 2;
-const int SETALL_HELL = 3;
-
-void setall_inspect() {
-	if (setall_inspect_list.size() == 0) {
-		return;
-	}
-	auto it = setall_inspect_list.begin();
-	CharData *ch = nullptr;
-	DescriptorData *d_vict = nullptr;
-
-	DescriptorData *imm_d = DescriptorByUid(player_table[it->first].uid());
-	if (!imm_d
-		|| (STATE(imm_d) != CON_PLAYING)
-		|| !(ch = imm_d->character.get())) {
-		setall_inspect_list.erase(it->first);
-		return;
-	}
-
-	timeval start{}, stop{}, result{};
-	int is_online;
-	gettimeofday(&start, nullptr);
-	Player *vict;
-	for (; it->second->pos < static_cast<int>(player_table.size()); it->second->pos++) {
-		vict = new Player;
-		gettimeofday(&stop, nullptr);
-		timediff(&result, &stop, &start);
-		if (result.tv_sec > 0 || result.tv_usec >= kOptUsec) {
-			delete vict;
-			return;
-		}
-		buf1[0] = '\0';
-		is_online = 0;
-		d_vict = DescriptorByUid(player_table[it->second->pos].uid());
-		if (d_vict)
-			is_online = 1;
-		if (player_table[it->second->pos].mail)
-			if (strstr(player_table[it->second->pos].mail, it->second->mail)) {
-				it->second->found++;
-				if (it->second->type_req == SETALL_FREEZE) {
-					if (is_online) {
-						if (GetRealLevel(d_vict->character) >= kLvlGod) {
-							sprintf(buf1, "Персонаж %s бессмертный!\r\n", player_table[it->second->pos].name());
-							it->second->out += buf1;
-							continue;
-						}
-						punishments::set_punish(imm_d->character.get(),
-								   d_vict->character.get(),
-								   SCMD_FREEZE,
-								   it->second->reason,
-								   it->second->freeze_time);
-					} else {
-						if (LoadPlayerCharacter(player_table[it->second->pos].name(), vict,
-												ELoadCharFlags::kFindId | ELoadCharFlags::kNoCrcCheck) < 0) {
-							sprintf(buf1, "Ошибка загрузки персонажа: %s.\r\n", player_table[it->second->pos].name());
-							delete vict;
-							it->second->out += buf1;
-							continue;
-						} else {
-							if (GetRealLevel(vict) >= kLvlGod) {
-								sprintf(buf1, "Персонаж %s бессмертный!\r\n", player_table[it->second->pos].name());
-								it->second->out += buf1;
-								continue;
-							}
-							punishments::set_punish(imm_d->character.get(),
-									   vict,
-									   SCMD_FREEZE,
-									   it->second->reason,
-									   it->second->freeze_time);
-							vict->save_char();
-						}
-					}
-				} else if (it->second->type_req == SETALL_EMAIL) {
-					if (is_online) {
-						if (GetRealLevel(d_vict->character) >= kLvlGod) {
-							sprintf(buf1, "Персонаж %s бессмертный!\r\n", player_table[it->second->pos].name());
-							it->second->out += buf1;
-							continue;
-						}
-						strncpy(GET_EMAIL(d_vict->character), it->second->newmail, 127);
-						*(GET_EMAIL(d_vict->character) + 127) = '\0';
-						sprintf(buf2,
-								"Смена e-mail адреса персонажа %s с %s на %s.\r\n",
-								player_table[it->second->pos].name(),
-								player_table[it->second->pos].mail,
-								it->second->newmail);
-						AddKarma(d_vict->character.get(), buf2, GET_NAME(imm_d->character));
-						it->second->out += buf2;
-
-					} else {
-						if (LoadPlayerCharacter(player_table[it->second->pos].name(), vict,
-												ELoadCharFlags::kFindId | ELoadCharFlags::kNoCrcCheck) < 0) {
-							sprintf(buf1, "Ошибка загрузки персонажа: %s.\r\n", player_table[it->second->pos].name());
-							it->second->out += buf1;
-							delete vict;
-							continue;
-						} else {
-							if (GetRealLevel(vict) >= kLvlGod) {
-								it->second->out += buf1;
-								continue;
-							}
-							strncpy(GET_EMAIL(vict), it->second->newmail, 127);
-							*(GET_EMAIL(vict) + 127) = '\0';
-							sprintf(buf2,
-									"Смена e-mail адреса персонажа %s с %s на %s.\r\n",
-									player_table[it->second->pos].name(),
-									player_table[it->second->pos].mail,
-									it->second->newmail);
-							it->second->out += buf2;
-							AddKarma(vict, buf2, GET_NAME(imm_d->character));
-							vict->save_char();
-						}
-					}
-				} else if (it->second->type_req == SETALL_PSWD) {
-					if (is_online) {
-						if (GetRealLevel(d_vict->character) >= kLvlGod) {
-							sprintf(buf1, "Персонаж %s бессмертный!\r\n", player_table[it->second->pos].name());
-							it->second->out += buf1;
-							continue;
-						}
-						Password::set_password(d_vict->character.get(), std::string(it->second->pwd));
-						sprintf(buf2, "У персонажа %s изменен пароль (setall).", player_table[it->second->pos].name());
-						it->second->out += buf2;
-						sprintf(buf1, "\r\n");
-						it->second->out += buf1;
-						AddKarma(d_vict->character.get(), buf2, GET_NAME(imm_d->character));
-					} else {
-						if (LoadPlayerCharacter(player_table[it->second->pos].name(), vict,
-												ELoadCharFlags::kFindId | ELoadCharFlags::kNoCrcCheck) < 0) {
-							sprintf(buf1, "Ошибка загрузки персонажа: %s.\r\n", player_table[it->second->pos].name());
-							it->second->out += buf1;
-							delete vict;
-							continue;
-						}
-						if (GetRealLevel(vict) >= kLvlGod) {
-							sprintf(buf1, "Персонаж %s бессмертный!\r\n", player_table[it->second->pos].name());
-							it->second->out += buf1;
-							continue;
-						}
-						Password::set_password(vict, std::string(it->second->pwd));
-						std::string str = player_table[it->second->pos].name();
-						str[0] = UPPER(str[0]);
-						sprintf(buf2, "У персонажа %s изменен пароль (setall).", player_table[it->second->pos].name());
-						it->second->out += buf2;
-						sprintf(buf1, "\r\n");
-						it->second->out += buf1;
-						AddKarma(vict, buf2, GET_NAME(imm_d->character));
-						vict->save_char();
-					}
-				} else if (it->second->type_req == SETALL_HELL) {
-					if (is_online) {
-						if (GetRealLevel(d_vict->character) >= kLvlGod) {
-							sprintf(buf1, "Персонаж %s бессмертный!\r\n", player_table[it->second->pos].name());
-							it->second->out += buf1;
-							continue;
-						}
-						punishments::set_punish(imm_d->character.get(),
-								   d_vict->character.get(),
-								   SCMD_HELL,
-								   it->second->reason,
-								   it->second->freeze_time);
-					} else {
-						if (LoadPlayerCharacter(player_table[it->second->pos].name(), vict,
-												ELoadCharFlags::kFindId | ELoadCharFlags::kNoCrcCheck) < 0) {
-							sprintf(buf1, "Ошибка загрузки персонажа: %s.\r\n", player_table[it->second->pos].name());
-							delete vict;
-							it->second->out += buf1;
-							continue;
-						} else {
-							if (GetRealLevel(vict) >= kLvlGod) {
-								sprintf(buf1, "Персонаж %s бессмертный!\r\n", player_table[it->second->pos].name());
-								it->second->out += buf1;
-								continue;
-							}
-							punishments::set_punish(imm_d->character.get(),
-									   vict,
-									   SCMD_HELL,
-									   it->second->reason,
-									   it->second->freeze_time);
-							vict->save_char();
-						}
-					}
-				}
-			}
-		delete vict;
-	}
-	if (it->second->mail && it->second->pwd)
-		Password::send_password(it->second->mail, it->second->pwd);
-	// освобождение памяти
-	if (it->second->pwd)
-		free(it->second->pwd);
-	if (it->second->reason)
-		free(it->second->reason);
-	if (it->second->newmail)
-		free(it->second->newmail);
-	if (it->second->mail)
-		free(it->second->mail);
-	gettimeofday(&stop, nullptr);
-	timediff(&result, &stop, &it->second->start);
-	sprintf(buf1, "Всего найдено: %d.\r\n", it->second->found);
-	it->second->out += buf1;
-	page_string(ch->desc, it->second->out);
-	setall_inspect_list.erase(it->first);
-}
-
-void do_setall(CharData *ch, char *argument, int/* cmd*/, int/* subcmd*/) {
-	int type_request = 0;
-	int times = 0;
-	if (ch->get_pfilepos() < 0)
-		return;
-
-	auto it = setall_inspect_list.find(GET_UID(ch));
-	// На всякий случай разрешаем только одну команду такого типа - либо setall, либо inspect
-	if (MUD::InspectRequests().IsBusy(ch) && it != setall_inspect_list.end()) {
-		SendMsgToChar(ch, "Обрабатывается другой запрос, подождите...\r\n");
-		return;
-	}
-
-	argument = three_arguments(argument, buf, buf1, buf2);
-	SetAllInspReqPtr req(new setall_inspect_request);
-	req->newmail = nullptr;
-	req->mail = nullptr;
-	req->reason = nullptr;
-	req->pwd = nullptr;
-
-	if (!*buf) {
-		SendMsgToChar("Usage: setall <e-mail> <email|passwd|frozen|hell> <arguments>\r\n", ch);
-		return;
-	}
-
-	if (!IsValidEmail(buf)) {
-		SendMsgToChar("Некорректный e-mail!\r\n", ch);
-		return;
-	}
-
-	if (!isname(buf1, "frozen email passwd hell")) {
-		SendMsgToChar("Данное действие совершить нельзя.\r\n", ch);
-		return;
-	}
-	if (utils::IsAbbr(buf1, "frozen")) {
-		skip_spaces(&argument);
-		if (!argument || !*argument) {
-			SendMsgToChar("Необходимо указать причину такой немилости.\r\n", ch);
-			return;
-		}
-		if (*buf2) times = atol(buf2);
-		type_request = SETALL_FREEZE;
-		req->freeze_time = times;
-		req->reason = strdup(argument);
-	} else if (utils::IsAbbr(buf1, "email")) {
-		if (!*buf2) {
-			SendMsgToChar("Укажите новый e-mail!\r\n", ch);
-			return;
-		}
-		if (!IsValidEmail(buf2)) {
-			SendMsgToChar("Новый e-mail некорректен!\r\n", ch);
-			return;
-		}
-		req->newmail = strdup(buf2);
-		type_request = SETALL_EMAIL;
-	} else if (utils::IsAbbr(buf1, "passwd")) {
-		if (!*buf2) {
-			SendMsgToChar("Укажите новый пароль!\r\n", ch);
-			return;
-		}
-		req->pwd = strdup(buf2);
-		type_request = SETALL_PSWD;
-	} else if (utils::IsAbbr(buf1, "hell")) {
-		skip_spaces(&argument);
-		if (!argument || !*argument) {
-			SendMsgToChar("Необходимо указать причину такой немилости.\r\n", ch);
-			return;
-		}
-		if (*buf2) times = atol(buf2);
-		type_request = SETALL_HELL;
-		req->freeze_time = times;
-		req->reason = strdup(argument);
-	} else {
-		SendMsgToChar("Какой-то баг. Вы эту надпись видеть не должны.\r\n", ch);
-		return;
-	}
-
-	req->type_req = type_request;
-	req->mail = str_dup(buf);
-	req->pos = 0;
-	req->found = 0;
-	req->out = "";
-	setall_inspect_list[ch->get_pfilepos()] = req;
 }
 
 void do_echo(CharData *ch, char *argument, int/* cmd*/, int subcmd) {
@@ -1707,76 +1408,6 @@ void do_restore(CharData *ch, char *argument, int/* cmd*/, int subcmd) {
 	}
 }
 
-void perform_immort_vis(CharData *ch) {
-	if (GET_INVIS_LEV(ch) == 0 &&
-		!AFF_FLAGGED(ch, EAffect::kHide) && !AFF_FLAGGED(ch, EAffect::kInvisible)
-		&& !AFF_FLAGGED(ch, EAffect::kDisguise)) {
-		SendMsgToChar("Ну вот вас и заметили. Стало ли вам легче от этого?\r\n", ch);
-		return;
-	}
-
-	SET_INVIS_LEV(ch, 0);
-	appear(ch);
-	SendMsgToChar("Вы теперь полностью видны.\r\n", ch);
-}
-
-void perform_immort_invis(CharData *ch, int level) {
-	if (ch->IsNpc()) {
-		return;
-	}
-
-	for (const auto tch : world[ch->in_room]->people) {
-		if (tch == ch) {
-			continue;
-		}
-
-		if (GetRealLevel(tch) >= GET_INVIS_LEV(ch) && GetRealLevel(tch) < level) {
-			act("Вы вздрогнули, когда $n растворил$u на ваших глазах.",
-				false, ch, nullptr, tch, kToVict);
-		}
-
-		if (GetRealLevel(tch) < GET_INVIS_LEV(ch) && GetRealLevel(tch) >= level) {
-			act("$n медленно появил$u из пустоты.",
-				false, ch, nullptr, tch, kToVict);
-		}
-	}
-
-	SET_INVIS_LEV(ch, level);
-	sprintf(buf, "Ваш уровень невидимости - %d.\r\n", level);
-	SendMsgToChar(buf, ch);
-}
-
-void do_invis(CharData *ch, char *argument, int/* cmd*/, int/* subcmd*/) {
-	int level;
-
-	if (ch->IsNpc()) {
-		SendMsgToChar("Вы не можете сделать этого.\r\n", ch);
-		return;
-	}
-
-	one_argument(argument, arg);
-	if (!*arg) {
-		if (GET_INVIS_LEV(ch) > 0)
-			perform_immort_vis(ch);
-		else {
-			if (GetRealLevel(ch) < kLvlImplementator)
-				perform_immort_invis(ch, kLvlImmortal);
-			else
-				perform_immort_invis(ch, GetRealLevel(ch));
-		}
-	} else {
-		level = MIN(atoi(arg), kLvlImplementator);
-		if (level > GetRealLevel(ch) && !ch->IsFlagged(EPrf::kCoderinfo))
-			SendMsgToChar("Вы не можете достичь невидимости выше вашего уровня.\r\n", ch);
-		else if (GetRealLevel(ch) < kLvlImplementator && level > kLvlImmortal && !ch->IsFlagged(EPrf::kCoderinfo))
-			perform_immort_invis(ch, kLvlImmortal);
-		else if (level < 1)
-			perform_immort_vis(ch);
-		else
-			perform_immort_invis(ch, level);
-	}
-}
-
 void do_gecho(CharData *ch, char *argument, int/* cmd*/, int/* subcmd*/) {
 	DescriptorData *pt;
 
@@ -1961,77 +1592,6 @@ void do_last(CharData *ch, char *argument, int/* cmd*/, int/* subcmd*/) {
 	}
 }
 
-void do_force(CharData *ch, char *argument, int/* cmd*/, int/* subcmd*/) {
-	DescriptorData *i, *next_desc;
-	char to_force[kMaxInputLength + 2];
-
-	half_chop(argument, arg, to_force);
-
-	sprintf(buf1, "$n принудил$g вас '%s'.", to_force);
-
-	if (!*arg || !*to_force) {
-		SendMsgToChar("Кого и что вы хотите принудить сделать?\r\n", ch);
-	} else if (!IS_GRGOD(ch) || (str_cmp("all", arg) && str_cmp("room", arg) && str_cmp("все", arg)
-		&& str_cmp("здесь", arg))) {
-		const auto vict = get_char_vis(ch, arg, EFind::kCharInWorld);
-		if (!vict) {
-			SendMsgToChar(NOPERSON, ch);
-		} else if (!vict->IsNpc() && GetRealLevel(ch) <= GetRealLevel(vict) && !ch->IsFlagged(EPrf::kCoderinfo)) {
-			SendMsgToChar("Господи, только не это!\r\n", ch);
-		} else {
-			char *pstr;
-			SendMsgToChar(OK, ch);
-			act(buf1, true, ch, nullptr, vict, kToVict);
-			sprintf(buf, "(GC) %s forced %s to %s", GET_NAME(ch), GET_NAME(vict), to_force);
-			while ((pstr = strchr(buf, '%')) != nullptr) {
-				pstr[0] = '*';
-			}
-			mudlog(buf, NRM, std::max(kLvlGod, GET_INVIS_LEV(ch)), SYSLOG, true);
-			imm_log("%s forced %s to %s", GET_NAME(ch), GET_NAME(vict), to_force);
-			command_interpreter(vict, to_force);
-		}
-	} else if (!str_cmp("room", arg)
-		|| !str_cmp("здесь", arg)) {
-		SendMsgToChar(OK, ch);
-		sprintf(buf, "(GC) %s forced room %d to %s", GET_NAME(ch), GET_ROOM_VNUM(ch->in_room), to_force);
-		mudlog(buf, NRM, std::max(kLvlGod, GET_INVIS_LEV(ch)), SYSLOG, true);
-		imm_log("%s forced room %d to %s", GET_NAME(ch), GET_ROOM_VNUM(ch->in_room), to_force);
-
-		const auto people_copy = world[ch->in_room]->people;
-		for (const auto vict : people_copy) {
-			if (!vict->IsNpc()
-				&& GetRealLevel(vict) >= GetRealLevel(ch)
-				&& !ch->IsFlagged(EPrf::kCoderinfo)) {
-				continue;
-			}
-
-			act(buf1, true, ch, nullptr, vict, kToVict);
-			command_interpreter(vict, to_force);
-		}
-	} else        // force all
-	{
-		SendMsgToChar(OK, ch);
-		sprintf(buf, "(GC) %s forced all to %s", GET_NAME(ch), to_force);
-		mudlog(buf, NRM, std::max(kLvlGod, GET_INVIS_LEV(ch)), SYSLOG, true);
-		imm_log("%s forced all to %s", GET_NAME(ch), to_force);
-
-		for (i = descriptor_list; i; i = next_desc) {
-			next_desc = i->next;
-
-			const auto vict = i->character;
-			if (STATE(i) != CON_PLAYING
-				|| !vict
-				|| (!vict->IsNpc() && GetRealLevel(vict) >= GetRealLevel(ch)
-					&& !ch->IsFlagged(EPrf::kCoderinfo))) {
-				continue;
-			}
-
-			act(buf1, true, ch, nullptr, vict.get(), kToVict);
-			command_interpreter(vict.get(), to_force);
-		}
-	}
-}
-
 void do_sdemigod(CharData *ch, char *argument, int/* cmd*/, int/* subcmd*/) {
 	DescriptorData *d;
 	// убираем пробелы
@@ -2060,132 +1620,6 @@ void do_sdemigod(CharData *ch, char *argument, int/* cmd*/, int/* subcmd*/) {
 				}
 			}
 		}
-	}
-}
-
-void do_wiznet(CharData *ch, char *argument, int/* cmd*/, int/* subcmd*/) {
-	DescriptorData *d;
-	char emote = false;
-	char bookmark1 = false;
-	char bookmark2 = false;
-	int level = kLvlGod;
-
-	skip_spaces(&argument);
-	delete_doubledollar(argument);
-
-	if (!*argument) {
-		SendMsgToChar
-			("Формат: wiznet <text> | #<level> <text> | *<emotetext> |\r\n "
-			 "        wiznet @<level> *<emotetext> | wiz @\r\n", ch);
-		return;
-	}
-
-	//	if (EPrf::FLAGGED(ch, EPrf::CODERINFO)) return;
-
-	// Опускаем level для gf_demigod
-	if (GET_GOD_FLAG(ch, EGf::kDemigod))
-		level = kLvlImmortal;
-
-	// использование доп. аргументов
-	switch (*argument) {
-		case '*': emote = true;
-			break;
-		case '#':
-			// Установить уровень имм канала
-			one_argument(argument + 1, buf1);
-			if (is_number(buf1)) {
-				half_chop(argument + 1, buf1, argument);
-				level = MAX(atoi(buf1), kLvlImmortal);
-				if (level > GetRealLevel(ch)) {
-					SendMsgToChar("Вы не можете изрекать выше вашего уровня.\r\n", ch);
-					return;
-				}
-			} else if (emote)
-				argument++;
-			break;
-		case '@':
-			// Обнаруживаем всех кто может (теоретически) нас услышать
-			for (d = descriptor_list; d; d = d->next) {
-				if (STATE(d) == CON_PLAYING &&
-					(IS_IMMORTAL(d->character) || GET_GOD_FLAG(d->character, EGf::kDemigod)) &&
-					!d->character->IsFlagged(EPrf::kNoWiz) && (CAN_SEE(ch, d->character) || IS_IMPL(ch))) {
-					if (!bookmark1) {
-						strcpy(buf1,
-							   "Боги/привилегированные которые смогут (наверное) вас услышать:\r\n");
-						bookmark1 = true;
-					}
-					sprintf(buf1 + strlen(buf1), "  %s", GET_NAME(d->character));
-					if (d->character->IsFlagged(EPlrFlag::kWriting))
-						strcat(buf1, " (пишет)\r\n");
-					else if (d->character->IsFlagged(EPlrFlag::kMailing))
-						strcat(buf1, " (пишет письмо)\r\n");
-					else
-						strcat(buf1, "\r\n");
-				}
-			}
-			for (d = descriptor_list; d; d = d->next) {
-				if (STATE(d) == CON_PLAYING &&
-					(IS_IMMORTAL(d->character) || GET_GOD_FLAG(d->character, EGf::kDemigod)) &&
-					d->character->IsFlagged(EPrf::kNoWiz) && CAN_SEE(ch, d->character)) {
-					if (!bookmark2) {
-						if (!bookmark1)
-							strcpy(buf1,
-								   "Боги/привилегированные которые не смогут вас услышать:\r\n");
-						else
-							strcat(buf1,
-								   "Боги/привилегированные которые не смогут вас услышать:\r\n");
-
-						bookmark2 = true;
-					}
-					sprintf(buf1 + strlen(buf1), "  %s\r\n", GET_NAME(d->character));
-				}
-			}
-			SendMsgToChar(buf1, ch);
-
-			return;
-		default: break;
-	}
-	if (ch->IsFlagged(EPrf::kNoWiz)) {
-		SendMsgToChar("Вы вне игры!\r\n", ch);
-		return;
-	}
-	skip_spaces(&argument);
-
-	if (!*argument) {
-		SendMsgToChar("Не думаю, что Боги одобрят это.\r\n", ch);
-		return;
-	}
-	if (level != kLvlGod) {
-		sprintf(buf1, "%s%s: <%d> %s%s\r\n", GET_NAME(ch),
-				emote ? "" : " богам", level, emote ? "<--- " : "", argument);
-	} else {
-		sprintf(buf1, "%s%s: %s%s\r\n", GET_NAME(ch), emote ? "" : " богам", emote ? "<--- " : "", argument);
-	}
-	snprintf(buf2, kMaxStringLength, "&c%s&n", buf1);
-	Remember::add_to_flaged_cont(Remember::wiznet_, buf2, level);
-
-	// пробегаемся по списку дескрипторов чаров и кто должен - тот услышит богов
-	for (d = descriptor_list; d; d = d->next) {
-		if ((STATE(d) == CON_PLAYING) &&    // персонаж должен быть в игре
-			((GetRealLevel(d->character) >= level) ||    // уровень равным или выше level
-				(GET_GOD_FLAG(d->character, EGf::kDemigod) && level == 31)) &&    // демигоды видят 31 канал
-			(!d->character->IsFlagged(EPrf::kNoWiz)) &&    // игрок с режимом NOWIZ не видит имм канала
-			(!d->character->IsFlagged(EPlrFlag::kWriting)) &&    // пишущий не видит имм канала
-			(!d->character->IsFlagged(EPlrFlag::kMailing)))    // отправляющий письмо не видит имм канала
-		{
-			// отправляем сообщение чару
-			snprintf(buf2, kMaxStringLength, "%s%s%s",
-					 CCCYN(d->character, C_NRM), buf1, CCNRM(d->character, C_NRM));
-			d->character->remember_add(buf2, Remember::ALL);
-			// не видино своих мессаг если 'режим repeat'
-			if (d != ch->desc
-				|| !(d->character->IsFlagged(EPrf::kNoRepeat))) {
-				SendMsgToChar(buf2, d->character.get());
-			}
-		}
-	}
-	if (ch->IsFlagged(EPrf::kNoRepeat)) {
-		SendMsgToChar(OK, ch);
 	}
 }
